Update patch set 6
Patch Set 6:

Hmm, the stats not being updated when pages are marked for translation seems to be related to the counting logic, not due to the the cache not being evicted. I tried doing null edits instead of usual purge but it doesn't get updated to match actual stats.
As for stats generated during normal translations, it does get correctly updated for me always locally so I think we should investigate further before changing that.

Patch-set: 6
Label: Verified=0