Clear WikiPageMessageGroup caches when page is marked for translation
MessageIndex and stats would go stale if JobQueue runs two
TranslationsUpdateJobs in same process, because then the
definitions would stay cached in the member variable of
WikiPageMessageGroup. While this kind of setup is uncommon,
most run jobqueue during web requests or cron, with the
--wait option or if lots of jobs are queue this can happen.

Test plan:
* Configure jobs to only be processed manually
* Run php maintenance/runJobs.php --wait
* Mark page for translation
* Do some translations for it
* Do changes to the page (add new messages)
* Mark the page for translation again

Without this patch the total number of messages would be stale.

Similar to T104172

Change-Id: Ib5b87f9807fb36a232800e08d3647395ebd0bb5c