Update patch set 1
Patch Set 1:

I would be happy to see both MessageGroupStats and MessageCollection to be refactored, but of course that is not a small undertaking so I cannot require that.

MessageGroupStats should be a class without public static methods and Database, MessageGroups and MessageCollection should be injected.

MessageCollection should be split into multiple classes: one for loading, one for filtering, and maybe one for expressing a collection of messages.

All the workarounds I can think are ugly, though, so I don't see how we could avoid refactorings.

Patch-set: 1