Update patch set 8
Patch Set 8: Code-Review-1

Still doesn't update stats.
Regarding the tests, I think it's fine even if it's not added in this patch if there is currently no set up for testing on translatable pages <--> translation units. We have already lost too much time on this and if this is fixed, admins would be a lot happier about this extension.

Patch-set: 8
Label: Code-Review=-1