braceBalanceCheck/XhtmlCheck: do not show warning if definition contains
Let's not enforce these checks even if the definition has warnable content as it is
counter-intuitive for translators. This is especially a problem for translatable
pages as the translation gets fuzzied if there is warning so the translation page
does not get re-rendered even if the unit was updated. That behavior should probably
be changed but that's for a separate patch.

Bug: T134692
Bug: T53226
Change-Id: I8943a0a9d50edeb82214c22106fc5e15cd91d1bf