Make deletions of translation units more performant
Currently, we are clearing the stats for all the languages for the
group and then recalculating them and inserting for all the languages.
That is hundreds of rows and requires doing the read query from
MessageCollection::loadInfo() a few hundred times. This is absolutely
insane while we are doing a deletion for just one single language.
So to increase the speed of this process, we clear the stats for only
the unit's language and the groups to which the unit belongs to. And then
recalculate the stats for only this language for the primary group.
Other groups can be calculated when someone views the stats and it
will get generated somewhere else.

Change-Id: I234de1d26bf3f705fe9c373302444a902094cd3a