Defer translation page update edit after translation unit deletion to a separate transaction
Users can use tools which can mass delete (eg. Nuke) multiple translation units
in one request. This means onDeleteTranslationUnit() will do many edits to a
translation page (and do MessageGroupStats queries) many times if many
translation units belonging to one translation page was deleted in one
request. This is quite bad user experience and can also be slow for the user.

To avoid this, we are now deferring the update to a separate transaction after
the main transaction so that we can use the text that is after all the units
has been deleted. Additionally, we now use a queue-based system so that only one update
is done for each translation page in order to avoid doing unnecessary stats updates,
null edits and purges (and other secondary updates) to these pages.

I'm not sure whether this will work in all cases where users can do multiple deletions
in a single request but at least this works for Nuke which is what is used on Wikimedia
for doing batch deletions in one request.
This was suggested at https://meta.wikimedia.org/?oldid=15647503#Deleting_translations

Change-Id: I2df5411fb51b6e45888fb7333855cf5a75732af9