Update patch set 1
Patch Set 1:

(1 comment)

Yes, Nuke does all the deletions in the main transaction which is committed before the callback here is executed. That means updateTranslationPage() will see the page text to be rendered after all the unit deletions has been done when using Nuke so we can do just one single edit with all the changes.

However since we're using a static variable for the queue, we cannot do a re-render in case a unit deletion is done after callback is executed and after being added to the queue previously. But since that's not the case for Nuke (which is the only case that I'm aware of where this can happen), that won't happen (or so I hope.. :)).

Patch-set: 1