Avoid fatal errors in TranslationsUpdateJob
The local message group process cache would become stale in long
running scripts such as runJobs.php --wait. Added a way to clear
the process cache and used it in the beginning of the job. Added
timestamp to the cached value to make debugging easier. Thanks to
it I realised that in my development environment I was using APC
for the message group cache. But APC is not shared with the web
requests and cli access, so I still got the failure because the
cli would store it's own cache which it would load again and again.
Fixed by switching to Database cache for my development environment.

Change-Id: I9e4834ef3c1eeb7343c2d37f59a93381853d974f