Allow accessing Special:Undelete for translation pages
This doesn't allow to do a "full" restoration if a translation page
was deleted along with its translation units but doing a full undeletion
doesn't seem like a trivial task. Since the units can also be restored
separately if needed, I think this is better than preventing
viewing and restoring deleted revisions completely.

Bug: T125339
Change-Id: Id90bafeabbac7161c7f971259d1051d8dad24f31