Always output a metadata block
People insist on having empty metadata appearing in the files.
I don't see the value in that, but apparently it is easier for
me to change the code than alter the test.

Or alternatively there is a misunderstanding what is the expected
behavior. We did always add a metadata block when we had authors to add.

Bug: T118205
Change-Id: I2e538b98e75da4fbc0ff960bd62ec9f508693afa