TranslateDeleteJob: Fix result of errors value when deletions fail
doDeleteArticle() returns a bool so it isn't very helpful to log this value
as an error. For some reason, it also seems to be either ignored by the
list=logevents API or it's not even inserted. So switch to use
doDeleteArticleReal() which returns a Status object so that we can properly
log the errors. Discovered while looking into T138571.

Change-Id: Ie04b7049accf5db4e79f136135d3c206e504e6a6