Update patch set 3
Patch Set 3: Code-Review+1

(3 comments)

Within the constraints discussed in the task, this looks like it should work. It relies on the wiki sysadmin to not use this with extensions that would return REDIRECT or UI responses during account creation, or extensions that would break somehow if accounts are deleted from the database.

Patch-set: 3
Label: Code-Review=+1