Update patch set 2
Patch Set 2:

The code in the special page still uses deprecated Title methods for checking whether the move is valid or not instead of using MovePage methods. I think it would make sense to use MovePage methods in this page before merging this. I can try to see whether that can be done (and other general cleanup in this page) before going forward with this. What do you think?

Patch-set: 2