ext.translate.editor: Improve save failure detection
mw.API() edit doesn't always return fail promises for actual
save failures. The proper way to check for success is checking
for the presence of "Success" in the API result. As we didn't check
for this previously, this caused some save failures to be reported
as successful to the user. Also now specific messages are shown for
AbuseFilter and SpamBlacklist save failures.

Bug: T142291
Bug: T61715
Change-Id: I7533d6aaee017aeb11f4c48658a35550afc94946