Update patch set 6
Patch Set 6:

I don't think it's safe to assume that it will happen quickly considering that this could be doing *lots* of moves (which is quite a heavy operation) even though it might be faster than the old code. Better to be on the safe side. I also don't it will be as complicated as it was previously because we are using just one job now.

Patch-set: 6
Label: Verified=0