Send proper Content-Type response header in Special:ExportTranslations
It used to be 'text/plain' but due to an oversight, that had been removed
and this now sends the default 'text/html' which is wrong here so change it
back again.

Also while at it, cleanup the code a little for consistency and use
WebResponse::header() instead calling header() directly.

Change-Id: Idd4e571ed416fa97e03211b9325f8e1db59eb1a6