tux: Change URL when the language is changed through language selector
Changing language client-side currently has several severe bugs and in
order to fix this in a quick non-hackish way, this commit changes the code
to change the URL instead of using client-side AJAX (incomplete) checks.
My plan is to stabilize the first load page with the upcoming major
refactoring of code executed for Special:Translate. Once we get to that
point, we can consider restoring the AJAX behavior once there are stable
and robust client-side interfaces for parameter changes.

As a consequence of this change, the "unsaved changes" dialog was shown
after our own dirty check dialog when a user tries to change the language.
I removed Translate's own check for this as browsers do not allow us to
reliably change this behavior and two confirmations for the same error is
poor form. This unfortunately means a custom MW-localized message cannot
be shown anymore. As I don't expect translators actively doing translation
would usually want to change the language in dirty state, the benefits
outweight the downsides here, IMO.

Bug: T51898
Bug: T134038
Bug: T142840
Change-Id: Ic772adf85a055cb12bf33df71720450d3d9817de