Fetch transclusion text from translation page instead of source page
Currently if a translatable page is transcluded, it shows the transclusion
with all the <translate> tags and all without any parser processing by Translate.
This commit changes it to fetch the transclusion text from the translation page
in the page language. If a translation does not exist in the page language,
then then text is retrieved from the source language page of the template.
Also adds a templatelinks dependency for the translation page.

Bug: T47096
Change-Id: Iec2bcc0417a16858c056ae75bd80aee28c62560c