Fetch transclusion text from translation page instead of source page
Currently if a translatable page is transcluded, it shows the transclusion
with all the <translate> tags and all without any parser processing by Translate.
This commit changes it to fetch the transclusion text from the translation page
in the page language. Also adds a templatelinks depency for the translation page.

Bug: T47096
Change-Id: Iec2bcc0417a16858c056ae75bd80aee28c62560c
Depends-On: I0de1c96bb968a99a2c81a9977655780a78988a20