Update patch set 2
Patch Set 2:

As I mentioned on a previous patch, I am planning to consolidate the language checking code to a single place but doing that in this patch would make it a bit more complicated than needed. It can be done in a follow-up.

Patch-set: 2
Reviewer: Glaisher <892@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0