TUX: Simplify message search results row
Using flexbox with justify-content: space-between to have the message
start from the left, and button "float" on the right. This greatly
simplifies the code and styles and makes more efficient use of space
in small screens. For browsers which do not support flexbox, the button
will show below the text, which I think is graceful degradation.

This fixes some existing issues such as of vertical alignment, huge size
of the advanced search button in mobile view, and color of the button.

Change-Id: I635fad84eda99b1fa4bb5321f17277e6d7d1a4a5