Support for CLDR plural keywords in JSON file format
This adds a configuration flag, parseCLDRPlurals, for the JSON
format. If set to yes, we will flatten/unflatten plural keywords
in the JSON source similarly to how it's already done for YAML,
using largely identical code (but also supporting the separator
option).

This makes redundant the YAML subclass implementation, which will
be refactored out in a follow-up change.

Change-Id: I15a9fb4a7f01f7ddb2f0b64c3a81062219dc2ac9