Implement plural aware message content comparison
When someone uses an inline plural in a messages that is exported
in a format that uses CLDR plural format without support for inline
syntax, on next sync that message is shown as changed. Since the
translation is in fact equal and we would like to keep the inline
version in the wiki, amend the comparison function to compare some
kind of canonical representation instead of plain string equality.

I moved this from ExternalMessageSourceStateComparator to FFS since
this only applies currently to few specific formats, and that there
might be other things that need similar handling in the future. For
example the fuzzy handling could also be moved to FFS classes.

Change-Id: I6417ce19332ee90346a00d5b658d45b39c3a691b