Implement plural aware message content comparison
When someone uses an inline plural in a message that is exported in a
format that uses the CLDR plural format without support for the inline
syntax, on the next import that message is shown as changed. Since the
translation is in fact equal and we would like to keep the inline
version in the wiki, amend the comparison function to compare a some
kind of canonical representation instead of using plain string equality.

I moved this code from ExternalMessageSourceStateComparator to FFS since
this only applies currently to a few specific formats, and because there
might be other things that need similar handling in the future. For
example the fuzzy handling could also be moved to the FFS classes.

Change-Id: I6417ce19332ee90346a00d5b658d45b39c3a691b