Remove $purpose parameter from password validity check
This was added in I56b6600 in an attempt to work around a bug in
CentralAuth, but the bug has since been fixed in a better way. No hook
functions in Gerrit use the parameter, and anything that was passing a
value other than the default 'login' has since been removed. So let's
just get rid of it instead of keeping it around doing nothing.

Change-Id: If7573198586804ade19e1e537f48e2834da99caf