Update patch set 1
Patch Set 1:

So given the module was renamed in core (but the old name is still available as b/c fallback), this should be renamed here too. But there should be also some conditional for MW <1.28.
I spent few minutes looking around on RL etc but have no good idea how to do it. Any suggestions/comments?

Change in Resources.php could be adjusted fairly easily so it checks for the MW version and uses a "right" module depending on the version. But what should happen to the JS code in ext.translate.editor.js then?

Patch-set: 1
Label: Verified=0