Update patch set 1
Patch Set 1:

> My first thought is that having <b> and </b> appear separately is
 > not what I would expect. I'm undecided whether this should go as-is
 > and be improved later, or improved bit more first.

As a user, especially as a translator to an RTL language, I would already appreciate separate <b> and </b> buttons very much as a considerable improvement.

Can be improved in further steps, but it's a good first one, and should be merged if the code is correct.

Patch-set: 1