Add support for Multi-DC for TTMServices
Introduce a new config key 'mirrors' that holds a list of TTM service
IDs. When a service defining mirrors is defined as the default TTM
service all writes received by TTMServerMessageUpdateJob will be
replicated to these mirrors.

TTMServerMessageUpdateJob now handles retries itself and will
reschedule a job to every single service that failed. The write
to a particular service is abandonned if too many failures occur.

When a TTMServerMessageUpdateJob is targeted at a specific service
mirroring is disabled.

If one of the mirror of the default service also defines mirrors the
job won't write to them. In other words we don't write to mirrors of
mirrors. This allows to set a configuration with 2 services mirroring
each others, enabling a fast switch of the default service.

Existing jobs in the jobqueue created with code prior to this patch
will be considered as initial jobs. Meaning that if the default service
has mirrors defined those jobs will also write to them.

If a job is targeted at an unknown service or a service that does
not implement WritableTTMServer it will be ignored.

When used with a JobQueue implementation based on redis resent jobs
are scheduled with a random backoff delay. The code to set the delay
was borrowed from the CirrusSearch extension.

Note that mirroring is not effective with the ttmserver-export.php script
as it deals directly with the service. The --ttmserver flag can be
used to initialize the mirror before activating it on live updates.

Bug: T132076
Change-Id: I4df596ac2444f1f988feb0fe9f0814d1413a6865