Add support for freezing writes
Add a new method isFrozen on WritableTTMServer. It allows services to
gracefully tell the update process that it does not accept writes.
Jobs attempting to write to a frozen service will be stored in the
jobqueue and retried multiple times. They will be abandoned after 1
day if the service remains frozen for a longer period.

It is usefull for maintainance tasks so that writes are not lost while
the service maintained.

Currently only ElasticTTMServer have been implemented by reusing the
Cirrus metastore. In order to freeze writes one must set the config
  "frozen_index" => "mw_cirrus_metastore"
and run the CirrusSearch maintenance script freezeWritesToCluster.php.

Bug: T132315
Change-Id: I6f15808672017769bae2c6c27e184ea7238eed2b