Add CaighdeanWebService
Not enabled by default.

While doing this I felt the urge to improve the exception handling
a bit: configuration or input errors should not count towards errors
that cause service to be suspended.

I also had to extend TranslationQuery and TranslationQueryResult to
allow passing some data, for implementing an oracle algorithm for
adding whitespace between tokens.

Change-Id: I7dc265f72c1965bcee4b916780706c52361876be