ElasticTTM: set the index when deleting docs
This was missed when migrating the code to elastic5. I thought that
setting the type would also set the index but elastica is sending a
query on _all/message/_search which is obviously extremely dangerous.
Fortunately this query was rejected on wmf production systems with the
error:
"Trying to query 3074 shards, which is over the limit of 1000."
Set the index on the Search object so that the request is no longer
using "_all" but ttmserver/message/_search.

Change-Id: Ifa46977d41472949f18d412a2803965994577bb7