Radio button for ignore vs. fuzzy on Special:ManageMessageGroups
With the checkboxes it's not clear what happens when both are selected.
A third radio button is needed to allow not to select either, but
no description is provided for the button since it's not clear what
happens in such a case (the translation administrator will have to
deduce by exclusion or personal knowledge, as is required now).

Bug: T160606
Change-Id: I201649c973a8b79efd201e84aaceee8b24f53608