Update patch set 3
Patch Set 3:

I asked and we made lots of discussions in #wikimedia-design. In short, Volker wants to know if the numbers are usually connected (making a spectrum) or they often can be found in groups close to each other. This is related to the comment made in the phab card. I would ask him to comment here or in the phab card

Patch-set: 3