Update patch set 3
Patch Set 3: Code-Review-1

(2 comments)

The old colors are what we call "coder colors": combinations of maxed out RGB ranges. Such colors are way to aggressive for almost everything you might want, except for error messages. And this here are not error messages, as far as I can tell.

The only problem I have with the proposed color scheme is that it is notably darker, and makes the black text harder to read. But this is just an opinion and not an edit request. See inline comments for the reason for my -1.

Patch-set: 3
Label: Code-Review=-1