Update patch set 2
Patch Set 2:

Both pieces are exactly the same as in EditPage.php, and I think they should remain so—the best would be to change really only the text of the button, but I don’t know how to do that. The current odd design is also because of the difference between the two pieces of code.

Patch-set: 2