Update patch set 1
Patch Set 1:

But you have to fix the translation before export to avoid breaking tests in the repository. Than there must be an exception on export with the message key and the exporteur has to rollback or delete the message key. Than he can restart the export. This sounds like a manual way which makes the export harder. Skipping the problematic translations does not change anything for the exporteur.

Another way is to skipping a whole group when there is at least one problematic translation, allow fixing it after export and than export the next day, could lead to an endless loop.

But having a self-merging export script which can break tests make it hard for other developer when it cames to this situation.

Patch-set: 1
Reviewer: Umherirrender <161@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0