Add showMessage and limit query parameters to Special:Translate
The purpose is to allow linking to specific messages in the message editor.
The lack of such feature is a blocker for removing the remaining legacy
non-TUX editor code. The code is implemented by manipulating the offset
and limit parameters in the message table. Some changes were needed
to make it not load more messages than requested. Changing any filter
will clear the showMessage parameter to resume normal use. The limit
parameter defaults to showing only one message, but it can be specified
to show multiple consecutive messages. The message editor is opened
automatically for the first message in this mode.

The exact format of the key required for the showMessage parameter
is hard to predict. It is basically the same as the key used in the
message group it originates from. Message groups take care of avoiding
invalid titles, but normalization of spaces and capitalization of
the first letter are still an issue.

For PHP code the format is not an issue, as it is easy to write a
function that returns the key that works for the offset. I am also
planning to add "Link to this message" feature to the editor. Use cases
such as linking from Special:Allmessages mostly work, due to the fact
that by convention most mediawiki messages start with a lower case
letter and don't use spaces.

Change-Id: I14af3d8746cb05d83e1e728e37723401ed30f5f4