[tech-debt] Remove usage of FuzzyLikeThis in favor of simple fuzzy match
It would allow us to remove a backport we maintain since elastic 1.x.
I've reviewed code of both queries and it should give the same results.

I'm still not 100% sure that it'll provide exactly the same results, I'd need
to have this code deployed and gather some feedback first.
Worst case scenario would be to revert this patch and keep maintaining the
FuzzyLikeThis query.
Best case would be removing hundreds of lines of code in the extra plugin.

This also removes the constraint to have the extra plugin installed to have
ttm working on top of elasticsearch (assuming you allow groovy scripts).

Change-Id: Id06e5d207c6bde1ff29e36ffbee509017632a97b