Fix editor shortcuts not showing up
JQMIGRATE: jQuery.fn.offset() requires an element connected to a document
Uncaught TypeError: $(...).text(...).offset(...).addClass is not a function

Tested positioning that was initially broken in Chrome, Firefox and Edge.
Seemed to work okay. Further testing is difficult as Alt press does not
go through in cross browser testing.

Change-Id: If7618ef0e82c7116f7acccf272ca604d0df4e834