Update patch set 2
Patch Set 2:

(4 comments)

As I've stated on the task, this is just to care about the most prominent icons and colors and a more aligning change would only make more sense with a thorough redesign of the extension as a whole. I don't see myself having time to regenerate the PNGs here. Given that the inconsistency affects subtle colors and browser like IE 9 and below I could live with it in contrast to the time needed to regenerate these.

Patch-set: 2