Fix log spam and exception with empty message definitions
Trying to query translation memory for a message which has an empty
definition (rare but happens) resulted in an uncaught exception
and few warnings. The logical thing is to just to skip all queries
if the definition is empty and that also avoids the reported issues.
This caused all translation aids to show "Loading..." indefinitely
for those messages.

Did the same for MachineTranslation aid. Even though it was not
causing visible issues, it seemed that the CXServer backend request
was internally timing out if the definition was empty.

Bug: T166104
Change-Id: I8698cabf08f16055c45fc2e27ace5b3af54d8a0a