2nd fix for TTMServer::sortSuggestions() must be of the type array, null given
Apart from empty strings as an input, a translation memory service
can fail in other ways too, for example because of a timeout.

TTMServerAid::formatSuggestions was not properly checking the
return value of TranslationWebService::getResultData.

Also correct return type for that method which can be anything what
a TranslationWebService subclass decides to return.

Tested manually by injecting a fault in RemoteTTMServerWebService
and observing that the error is gone.

Bug: T166104
Change-Id: I3152ca2ea25c1869ead3a80ce4eeb6c6c27b2a42