Drop `-webkit-linear-gradient` image CSS fallback
This reduces the payload of CSS slightly (and in the few cases where these
were not just present but @embed'ed, significantly), at a cost of giving
those user agents which don't support the un-prefixed rule the icons as
PNGs rather than SVGs (i.e. Chrome < 25, Safari < 6.1, Android < 4.4).

We made this change for MediaWiki back in November 2015 in T118815. This
repo is not alone in not having noticed. :-) Ideally this repo would use
LESS and re-use the LESS macro in MediaWiki, but that's for the future.

Change-Id: I1469bce3dd6c6eed0d291a0cb8a19eceeaff4c10