Add a script that expands message group spec
Specification is a comma separated list of message groups ids
which allows wildcards * and ?.

I am planning to make RepoNG to take advantage of parallelism.
In twn config we have groupspec such as mwgithub-* and RepoNG
cannot parallelize exports if it doesn't know the full list
of message groups that wildcard matches. This script provides
that.

Change-Id: I2bc6be824ccd5aa19e2ee15378150952f82e33a5