Update patch set 2
Patch Set 2:

I am not sure if that check is needed.
The problem is that if title is made not translatable, and loop is iterated starting from 1, then the first translation unit is not displayed, which is not title, but part of the text.
If I am not right about this - please correct me...

I wonder why such trivial bug is that old too, but I think that is because PageMigration is not actively used and this bug could remain unnoticed for a long time as nobody spotted absence of page title translation field.

Patch-set: 2