Update patch set 7
Patch Set 7:

Whoops I accidentally added my fuzzy script changes to PS7.

Basically this add --user option to the script to fuzzy translation of a particular user. The code has been tested in production, so it should be fine, even if the code is not as polished as usual.

Patch-set: 7