Update patch set 1
Patch Set 1:

Thanks for the comment.

Here's an example of a whole entry in the database:

| 396640 | ceb09f66148359b9804b94cc36dece3f | NULL | 20180321063557 | {"wmf_app_version": "-", "os_minor": "10", "os_major": "10", "is_bot": false, "device_family": "Other", "os_family": "Mac OS X", "browser_minor": "0", "is_mediawiki": false, "browser_major": "64", "browser_family": "Chrome"}     | test.wikipedia.org | testwiki      | no-search-results | NULL              | en                    | espano        | NULL              | NULL                | en                      | NULL                          | Special:ContentTranslation | CENSORED         |                        |             1 |

This is less than 700 bytes. In JSON format it will probably be different, but not by far.

It's good to be careful about it, but I don't think that we're at a point where we have to be extra-prudent yet.

We may consider reducing the schema fields, though, because we aren't actually using all of them.

Patch-set: 1
Reviewer: Amire80 <54@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0