Fix incorrect API param defaults
In two cases, the type was wrong (string vs. integer).

In two others, the allowed values were dynamically computed and didn't
always include the default value.  In these cases, the default was
removed entirely and the parameter is now required.

In the last case, a limit was incorrectly specified as being of type
'integer' rather than 'limit', which meant (among other things) that
PARAM_MAX2 was ignored.

Change-Id: Ied07ccfd7ff6db9d399eac3c8aab63993b91c1f4