Update patch set 3
Patch Set 3:

Sorry I didn't capture the exception, but it is in Translate code which tries to do stuff with the null parameter. It would be better if the API didn't let the code proceed without a parameter.

I guess the modules could try to disable themselves completely if there are no valid parameters, but that seems more involved.

Patch-set: 3