Optimize AggregateMessageGroup::getKeys
This started from observation that createMessageIndex.php took ~60
seconds when it should only take a few. Initial guess was that the
dba php extension was missing. It was, but installing it only made
it little bit slower, not faster (not a scientific evaluation).

Further investigation with xdebug + kcachegrind revealed that most
of the time (75 %) was spent on array_merge. Replacing that with a
append loop fixes that in a way that nothing particular sticks out
anymore. Array_unique takes a bit more time now but not much. Also
added array_keys to renumber the keys for a good measure.

I suspect the migration from HHVM to PHP made this bottleneck more
apparent, or that HHVM has a better implementation of array_merge.

Bug: T190685
Change-Id: I793b379459325f47f49a35e15ee1be6544502751