Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5: Code-Review-1
> > 
> > I'm not sure what happened here – now there is no default value at all. That would be a breaking change I don't want to do.
> 
> Currently it is both required, and has a default value.  This doesn't make sense: the default value is used if the caller doesn't provide a value, but required means that the caller must provide a value.

Er, sorry.  Submitted prematurely.  Let me try again:

You asked in your last review that the parameters be required.  Being required means having no default -- you can't have it be required and also have a default.  Currently this is allowed, but that's a bug, and it will be no longer allowed after I4a2505cfc8e86e73e46d40acffbf3cd8803458e2 lands.  If a default is specified, PARAM_REQUIRED is ignored.

So what do you actually want to happen if the client does not specify a service, and the default service is not in the list of allowed services?  What value should be used?  If you want $wgTranslateTranslationDefaultService to be used in this case, why shouldn't it be added to $available as I had in my first patch?

Patch-set: 5
Label: Verified=0