Update patch set 5
Patch Set 5:

> So what do you actually want to happen if the client does not specify a service, and the default service is not in the list of allowed services?

The API must fail gracefully with an error (not with an exception).

There are differents cases:
1) No (public) service is configured: the API is not usable. I am open to suggestions how to handle this better. This is likely the most common case as it is the default configuration.
2) There is one (public) service configured: There is only one possible value, and it is kind of nice to provide it as the default, because the value can vary. This is the second most common case.
3) There are multiple (public) services configured: There are multiple possible values and the client should be able to choose. One must still be default (not in the API sense but in terms of Translate configuration) and it makes sense to use it as the default value for the API too.

In this case I interpret PARAM_REQUIRED as follows: there must be a value, but the default value is also a valid value.
It seems you argue for the reading that PARAM_REQUIRED means: client must provide a value.

So to me it doesn't matter that PARAM_REQUIRED is ignored, as you say, if default value is provided, that's exactly what I want as a developer: to have a contract with the API parameter validation layer that my code is not executed unless a valid value is given.

I don't require the exact solution I proposed, but I do not want exceptions. What solution do you suggest?

Patch-set: 5